I need to make this update now because my private version is getting too different from the svn and the updates are not applying properly. I think i got all of the last updates manually applied, but someone might want to check
- update to the physics system for comments
- some text changes to fix spelling mistakes, some grammar problems, and some editing
- renamed the font to Bo_Chen, so we can possibly have multiple ttf fonts, rather than just font.ttf
- renamed some functions for consistency, etc.
- a couple of mods to some md2s
- some signed/unsigned warnings removed
- a bunch of other stuff I forgot

+++ game/bsp.h
@@ -179,7 +179,7 @@ BSP_branch_t * BSP_tree_ensure_branch( BSP_tree_t * t, BSP_branch_t * B, int ind
 Sint32         BSP_tree_count_nodes( Sint32 dim, Sint32 depth );
 bool_t         BSP_tree_insert( BSP_tree_t * t, BSP_branch_t * B, BSP_leaf_t * n, int index );
 bool_t         BSP_tree_insert_leaf( BSP_tree_t * ptree, BSP_leaf_t * pleaf );
-bool_t         BSP_tree_prune_branch( BSP_tree_t * t, int cnt );
+bool_t         BSP_tree_prune_branch( BSP_tree_t * t, size_t cnt );
 
 //--------------------------------------------------------------------------------------------
 //--------------------------------------------------------------------------------------------
